+++ Common/ListenerCatcher.cs
@@ -1,0 +1,7 @@
+﻿namespace DSystem
+{
+    public class ListenerCatcher<T> where T : class
+    {
+        public T Listener { get; internal set; }
+    }
+}